test: modify test after contract split
